refactor: ProductRequest와 OptionRequest로 AddProductRequest 구성
