+++ src/main/java/gift/dto/request/ProductRequest.java
@@ -1,0 +1,48 @@
+package gift.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.Size;
+
+import static gift.constant.Message.*;
+
+public class ProductRequest {
+
+    @NotBlank(message = REQUIRED_FIELD_MSG)
+    @Size(max = 15, message = LENGTH_ERROR_MSG)
+    private String name;
+
+    @NotNull(message = REQUIRED_FIELD_MSG)
+    @Positive(message = POSITIVE_NUMBER_REQUIRED_MSG)
+    private int price;
+
+    @NotBlank(message = REQUIRED_FIELD_MSG)
+    private String imageUrl;
+
+    @NotBlank(message = REQUIRED_FIELD_MSG)
+    private String category;
+
+    public ProductRequest(String name, int price, String imageUrl, String category) {
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+        this.category = category;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+}
